requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.2.14
langchain-community>=0.2.12
langchain-text-splitters>=0.2.4
langchain-core>=0.2.38
langchain-experimental>=0.0.65
faiss-cpu>=1.8.0
pydantic>=2.7.0
python-multipart>=0.0.9
pypdf>=4.3.1
tiktoken>=0.7.0
python-dotenv>=1.0.1
